requirements.txt
torch
transformers
einops
Pillow
sqlite3
langchain
langchain-openai
streamlit